third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-preserveAlpha-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAlpha' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('preserveAlpha') is "false"
PASS convolveMatrixElement.getAttribute('preserveAlpha') is "true"
PASS successfullyParsed is true

TEST COMPLETE
